+++ lbps/device.py
@@ -56,7 +56,7 @@ class Device(Bearer):
 		pre = "%s::capacity\t\t" % self._name
 
 		if self._CQI:
-			return N_TTI_RE*T_CQI[self._CQI]['eff']
+			return int(N_TTI_RE*T_CQI[self._CQI]['eff'])
 
 		msg_fail("failed", pre=pre)
 		return
@@ -66,7 +66,7 @@ class Device(Bearer):
 		pre = "%s::virtualCapacity\t" % self._name
 
 		if self.tdd_config:
-			return self.tdd_config.count('D')*self.capacity/10
+			return int(self.tdd_config.count('D')*self.capacity/10)
 
 		msg_fail("failed", pre=pre)
 		return
@@ -218,9 +218,9 @@ class RN(Device):
 
 		if self.tdd_config and self.__parent.idle_capacity:
 			times = self.tdd_config.count('D')+self.__parent.idle_capacity
-			return times*self.capacity['access']/10
+			return int(times*self.capacity['access']/10)
 		elif self.tdd_config:
-			return self.tdd_config.count('D')*self.capacity['access']/10
+			return int(self.tdd_config.count('D')*self.capacity['access']/10)
 
 		msg_fail("failed", pre=pre)
 		return
